requirements.txt
# RAG and Vector store
chromadb==0.5.17
sentence-transformers==3.3.1
PyPDF2==3.0.1
pypdf==5.1.0

# API and ML Utils
google-generativeai==0.8.3
scikit-learn==1.5.2
pandas==2.2.3

# Evaluation Metrics
rouge-score==0.1.2
nltk==3.9.1
